# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverProfileRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "rating_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Avg"
          },
          "status": {
            "$ref": "#/components/schemas/DriverStatus"
          },
          "truck": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TruckRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "truck_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truck Photo Url"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "user_id",
          "status",
          "verified",
          "license_url",
          "truck_photo_url",
          "rating_avg"
        ],
        "title": "DriverProfileRead",
        "type": "object"
      },
      "DriverRegisterRequest": {
        "description": "POST /v1/drivers/me/register body (AUTH-5, `invite_token` added by FLT-4).\n\nDocument upload to object storage is out of scope for this task \u2014 license/truck\nphoto URLs are accepted as plain strings for now; a later task wires real upload\nand replaces these with signed URLs from that flow.\n\nTwo mutually exclusive shapes: bring your own truck (`plate`/`truck_type`/\n`capacity`, the original AUTH-5 flow -- all three required), or redeem a fleet\nowner's invite (`invite_token` from POST /v1/fleets/me/invites), which already\npre-provisioned the truck -- `plate`/`truck_type`/`capacity` are ignored (and\nmust be left null) in that case. The endpoint 422s if both shapes are mixed.",
        "properties": {
          "capacity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TruckCapacity"
              },
              {
                "type": "null"
              }
            ]
          },
          "invite_token": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Token"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "plate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plate"
          },
          "truck_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truck Photo Url"
          },
          "truck_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TruckType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "DriverRegisterRequest",
        "type": "object"
      },
...
      "FleetBalanceRead": {
        "description": "Consolidated balance across every driver whose truck belongs to the fleet.",
        "properties": {
          "fleet_id": {
            "format": "uuid",
            "title": "Fleet Id",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/FleetMemberBalance"
            },
            "title": "Members",
            "type": "array"
          },
          "owed_balance": {
            "title": "Owed Balance",
            "type": "integer"
          }
        },
        "required": [
          "fleet_id",
          "owed_balance",
          "members"
        ],
        "title": "FleetBalanceRead",
        "type": "object"
      },
      "FleetCreate": {
        "description": "POST /v1/fleets/me body.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "FleetCreate",
        "type": "object"
      },
...
      "FleetRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "trucks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TruckRead"
            },
            "title": "Trucks",
            "type": "array"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "name",
          "created_at"
        ],
        "title": "FleetRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InviteCreate": {
        "description": "POST /v1/fleets/me/invites body.\n\nMirrors DriverRegisterRequest's truck fields exactly (same types, reused from\napp/schemas/driver.py) -- this pre-provisions the same Truck row AUTH-5's own\nregistration would otherwise create, just with no driver_id yet.",
        "properties": {
          "capacity": {
            "$ref": "#/components/schemas/TruckCapacity"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "plate": {
            "title": "Plate",
            "type": "string"
          },
          "truck_type": {
            "$ref": "#/components/schemas/TruckType"
          }
        },
        "required": [
          "phone",
          "plate",
          "truck_type",
          "capacity"
        ],
        "title": "InviteCreate",
        "type": "object"
      },
      "InviteRead": {
        "description": "Returned by POST /v1/fleets/me/invites and GET /v1/fleets/me/invites.\n\n`invite_token` is what the invited driver passes as `invite_token` on\nPOST /v1/drivers/me/register to redeem it.",
        "properties": {
          "invite_token": {
            "format": "uuid",
            "title": "Invite Token",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "truck_id": {
            "format": "uuid",
            "title": "Truck Id",
            "type": "string"
          }
        },
        "required": [
          "invite_token",
          "truck_id",
          "phone"
        ],
        "title": "InviteRead",
        "type": "object"
      },
...
      "TruckCapacity": {
        "enum": [
          "moto",
          "car",
          "both"
        ],
        "title": "TruckCapacity",
        "type": "string"
      },
      "TruckRead": {
        "properties": {
          "capacity": {
            "$ref": "#/components/schemas/TruckCapacity"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "driver_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "fleet_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fleet Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "plate": {
            "title": "Plate",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TruckType"
          }
        },
        "required": [
          "id",
          "plate",
          "type",
          "capacity",
          "driver_id",
          "fleet_id"
        ],
        "title": "TruckRead",
        "type": "object"
      },
...
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "fcm_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcm Token"
          },
          "firebase_uid": {
            "title": "Firebase Uid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "firebase_uid",
          "role",
          "name",
          "phone",
          "email",
          "fcm_token",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "The Crane API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/drivers/me/register": {
      "post": {
        "description": "AUTH-5: create driver_profiles (unverified, offline) + trucks, flip role -> driver.\n\nTwo mutually exclusive shapes (see DriverRegisterRequest's docstring): bring your\nown truck (`plate`/`truck_type`/`capacity`), or redeem a fleet owner's invite\n(`invite_token`, FLT-4) which already pre-provisioned the truck \u2014 mixing the two\nis 422. Redeeming an invite additionally requires the caller's verified Firebase\nphone claim to match the phone the invite was created for (403 otherwise), and\nmarks the invite consumed so it can't be redeemed twice (409 if it already was,\n404 if the token doesn't exist).\n\n409 if the caller is already a driver. Document upload is out of scope (see the\nschema docstring); license_url/truck_photo_url are accepted as-is.",
        "operationId": "register_driver_v1_drivers_me_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Driver",
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/v1/fleets/me": {
      "get": {
        "description": "The caller's fleet plus every truck currently attached to it.",
        "operationId": "get_my_fleet_v1_fleets_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Fleet",
        "tags": [
          "fleets"
        ]
      },
      "post": {
        "description": "FLT-1: create the caller's fleet and flip role -> fleet_owner.\n\n409 if the caller already owns a fleet (one fleet per owner, same convention as\ndriver_profiles' unique user_id).",
        "operationId": "create_fleet_v1_fleets_me_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Fleet",
        "tags": [
          "fleets"
        ]
      }
    },
    "/v1/fleets/me/balance": {
      "get": {
        "description": "FLT-2: consolidated owed balance across every driver in the caller's fleet,\nplus the per-driver breakdown it's built from (app/services/ledger.py).",
        "operationId": "get_my_fleet_balance_v1_fleets_me_balance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetBalanceRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Fleet Balance",
        "tags": [
          "fleets"
        ]
      }
    },
    "/v1/fleets/me/invites": {
      "get": {
        "description": "Pending invites the caller's fleet has outstanding (not yet redeemed).",
        "operationId": "list_my_invites_v1_fleets_me_invites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InviteRead"
                  },
                  "title": "Response List My Invites V1 Fleets Me Invites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Invites",
        "tags": [
          "fleets"
        ]
      },
      "post": {
        "description": "FLT-4: invite a driver who doesn't have a truck (or an account) yet.\n\nPre-provisions the Truck row up front (unclaimed by a driver, `fleet_id` = the\ncaller's fleet) and hands back a token; the invited driver redeems it via\n`invite_token` on POST /v1/drivers/me/register, which links them onto this exact\ntruck instead of creating a new one.\n\n409 if `phone` already has a pending invite (redeem or let it be superseded\nfirst), or if `plate` is already taken \u2014 mirrors AUTH-5's plate-uniqueness\nIntegrityError handling in app/api/drivers.py's register_driver.",
        "operationId": "create_invite_v1_fleets_me_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Invite",
        "tags": [
          "fleets"
        ]
      }
    },
    "/v1/fleets/me/trucks/{truck_id}": {
      "delete": {
        "description": "Detach a truck from the caller's fleet. 404 if it isn't currently a member.",
        "operationId": "remove_truck_from_fleet_v1_fleets_me_trucks__truck_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "truck_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Truck Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Truck From Fleet",
        "tags": [
          "fleets"
        ]
      },
      "post": {
        "description": "Attach an unclaimed truck (fleet_id is null) to the caller's fleet.\n\n404 if the caller has no fleet or the truck doesn't exist; 409 if the truck\nalready belongs to a fleet (its own or someone else's \u2014 detach it first).",
        "operationId": "add_truck_to_fleet_v1_fleets_me_trucks__truck_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "truck_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Truck Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Truck To Fleet",
        "tags": [
          "fleets"
        ]
      }
    },
    "/v1/fleets/trucks/by-plate/{plate}": {
      "get": {
        "description": "FLT-4: look up a truck before attaching it -- a fleet owner knows a driver's\nplate, not their truck's UUID. Any authenticated user may look up any plate\n(plates aren't secret; this doesn't expose anything `add_truck_to_fleet` below\nwouldn't already reveal via its 404/409 either way) -- callers only get to act\non the result via the existing per-fleet attach endpoint's own checks.",
        "operationId": "find_truck_by_plate_v1_fleets_trucks_by_plate__plate__get",
        "parameters": [
          {
            "in": "path",
            "name": "plate",
            "required": true,
            "schema": {
              "title": "Plate",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Find Truck By Plate",
        "tags": [
          "fleets"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "description": "Return the authenticated user's profile.",
        "operationId": "read_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Me",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Update name/email/fcm_token; fields absent from the body are left untouched.",
        "operationId": "update_me_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}